rpi-wm8804-soundcard: use nicer driver_name "RPi-WM8804"
The driver_name shows up in /proc/asound/devices and is used
to load the proper Alsa card conf from /usr/share/alsa/cards.

But the length is limited to 15 characters and spaces are
replaced by underscore so the mangled name currently is
"RPI_WM8804_soun" which is not very nice.

Use "RPi-WM8804" instead which won't be mangled and can be used
unaltered as the card conf filename and card name in that file.

Signed-off-by: Matthias Reichl <hias@horus.com>
(cherry picked from commit 4d5572813b386425ef0252290c961f8a55b82c64)
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>